tasks/requirements.txt
Flask==2.2.2
flask-sqlalchemy
hazelcast-python-client
